requirements.txt
aiohttp==3.8.3
aiohttp-cors==0.7.0
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
antlr4-python3-runtime==4.9.3
anyio==3.4.0
appdirs==1.4.4
argcomplete==1.12.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
async-timeout==4.0.2
asyncssh==2.13.1
atpublic==3.1.1
attrs==22.2.0
Babel==2.9.1
backcall==0.2.0
backoff==1.11.1
backports.zoneinfo==0.2.1
bcrypt==4.0.1
bidict==0.21.4
billiard==4.1.0
bleach==4.1.0
blessed==1.20.0
boto3==1.16.34
botocore==1.19.63
cachetools==5.0.0
captum==0.6.0
celery==5.3.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.0.0
colorama==0.4.6
colorful==0.5.5
compdb==0.2.0
configobj==5.0.8
contextlib2==21.6.0
contourpy==1.0.7
cryptography==36.0.1
cycler==0.11.0
dask==2023.5.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dictdiffer==0.9.0
dill==0.3.6
diskcache==5.4.0
distlib==0.3.7
distributed==2023.5.0
distro==1.8.0
docker==4.4.4
dpath==2.1.4
dulwich==0.21.6
dvc==3.18.0
dvc-data==2.15.4
dvc-http==2.30.2
dvc-objects==1.0.1
dvc-render==0.3.1
dvc-ssh==2.22.1
dvc-studio-client==0.15.0
dvc-task==0.3.0
dvclive==2.0.2
entrypoints==0.3
et-xmlfile==1.1.0
executing==1.2.0
ffi-navigator==0.5.4
filelock==3.9.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-SocketIO==5.1.1
flatbuffers==2.0
flatten-dict==0.4.2
flufl.lock==6.0
fonttools==4.29.1
frozenlist==1.3.3
fsspec==2023.1.0
funcy==1.18
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.11.1
google-auth==2.22.0
googleapis-common-protos==1.60.0
gpustat==1.1.1
grandalf==0.8
graphviz==0.20.1
greenlet==1.1.2
grpcio==1.58.0
hydra-core==1.3.1
hyperopt==0.2.7
idna==3.3
importlib-metadata==6.0.0
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==6.17.1
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==7.6.6
iterative-telemetry==0.0.7
itsdangerous==2.1.0
jedi==0.18.2
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
json5==0.9.6
jsonschema==4.3.2
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.6.3
jupyter-server==1.13.5
jupyter_core==5.3.1
jupyterlab==3.3.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
jupyterlab_server==2.16.3
kiwisolver==1.3.2
kombu==5.3.2
kubernetes==12.0.1
locket==1.0.0
lpips==0.1.4
markdown-it-py==2.1.0
MarkupSafe==2.0.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==0.8.4
modin==0.23.1
modin-spreadsheet==0.1.2
mpi4py-mpich==3.1.2
msgpack==1.0.5
multidict==6.0.4
nanotime==0.5.2
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
neural-compressor==1.9.1
notebook==6.4.6
numpy==1.21.4
nvidia-ml-py==12.535.108
oauthlib==3.2.0
omegaconf==2.3.0
onnx==1.14.1
onnxruntime==1.10.0
opencensus==0.11.2
opencensus-context==0.1.3
opencv-python==4.5.5.62
openpyxl==3.1.2
orjson==3.8.5
packaging==21.3
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.0
pathspec==0.11.0
patsy==0.5.6
PeachPy @ git+https://github.com/Maratyszcza/PeachPy@349e8f836142b2ed0efeb6bb99b1b715d87202e9
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
Pint==0.16.1
platformdirs==3.2.0
pluggy==1.0.0
prettytable==3.1.1
prometheus-client==0.12.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.8.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==8.0.0
py-spy==0.3.14
py4j==0.10.9.3
pyarrow==13.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.5
pycparser==2.21
pydantic==1.10.12
pydot==1.4.2
pyflann-py3==0.1.0
pyfunctional==1.4.3
pygit2==1.11.1
Pygments==2.11.0
pygtrie==2.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.6
pypng==0.0.21
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
python-engineio==4.3.1
python-jsonrpc-server==0.4.0
python-socketio==5.5.2
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.4.4
QtPy==2.4.0
ray==2.6.3
requests==2.26.0
requests-oauthlib==1.3.1
rich==13.2.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.3.7
schema==0.7.5
scikit-learn==1.0.2
scipy==1.7.3
scmrepo==1.3.1
seaborn==0.12.2
Send2Trash==1.8.0
shortuuid==1.0.11
shtab==1.5.8
sigopt==8.2.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqltrie==0.7.0
sshfs==2023.4.1
stack-data==0.6.2
statsmodels==0.14.1
tabulate==0.9.0
tblib==2.0.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.1.0
toml==0.10.2
tomlkit==0.11.6
toolz==0.12.0
# torch==1.13.1+cu117
# torchaudio==0.13.1+cu117
# torchvision==0.14.1+cu117
tornado==6.1
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.7.0
unidist==0.4.1
urllib3==1.26.7
vine==5.0.0
virtualenv==20.21.0
voluptuous==0.13.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.3
widgetsnbextension==3.5.2
xgboost==1.5.1
xmltodict==0.12.0
yarl==1.8.2
yq==2.13.0
zc.lockfile==2.0
zict==3.0.0
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0

captum
graphviz
seaborn
pyfunctional
dvc[ssh]==3.18.0
onnx==1.14.1
onnxruntime==1.10.0
ffi-navigator
lpips==0.1.4
ghidra-stubs==11.1.2.1.0.4
